Serialize Date objects to BSON:
For dates, we create a UTC time at midnight for the Date, and then
serialize the time.

There is no deserialization into Dates, since BSON only has a Time type,
but is useful for queries where a user is working with Date objects and
doesn't want to manually convert to a Time themselves at each instance.